Fix: CreateRoom.test.tsx 테스트 오류 수정
- setup.ts에 createRoomId함수를 모킹
- MakeRoom.test.tsx를 CreateRoom으로 이름변경
- setup.ts에 useParams도 모킹, ControlGame.test.tsx에서 useParams를 추가
  모킹하여 컴포넌트 오류 수정
- playSolo 테스트 컴포넌트 폴더 오타 수정
