+++ __test__/components/loby/playSolo/PlaySolo.test.tsx
