+++ package.json
@@ -34,6 +34,7 @@
     "zustand": "^4.5.5"
   },
   "devDependencies": {
+    "@faker-js/faker": "^9.1.0",
     "@mswjs/data": "^0.16.2",
     "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^16.0.0",
